backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
torch==2.1.0
pytorch-forecasting==1.0.0
lightning==2.1.0
pandas==2.1.3
numpy==1.24.3
yfinance==0.2.32
ta==0.11.0
bokeh==3.3.0
scikit-learn==1.3.2
python-dateutil==2.8.2
